# json_data/openapi.yaml
# VERITAS: Copyright (c) 2020 Veritas Technologies LLC. All rights reserved.
#
# THIS SOFTWARE CONTAINS CONFIDENTIAL INFORMATION AND TRADE SECRETS OF VERITAS
# TECHNOLOGIES LLC.  USE, DISCLOSURE OR REPRODUCTION IS PROHIBITED WITHOUT THE
# PRIOR EXPRESS WRITTEN PERMISSION OF VERITAS TECHNOLOGIES LLC.
#
# The Licensed Software and Documentation are deemed to be commercial computer
# software as defined in FAR 12.212 and subject to restricted rights as defined
# in FAR Section 52.227-19 "Commercial Computer Software - Restricted Rights"
# and DFARS 227.7202, Rights in "Commercial Computer Software or Commercial
# Computer Software Documentation," as applicable, and any successor
# regulations, whether delivered by Veritas as on premises or hosted services.
# Any use, modification, reproduction release, performance, display or
# disclosure of the Licensed Software and Documentation by the U.S. Government
# shall be solely in accordance with the terms of this Agreement.
# Product version __version__

openapi: 3.0.0
info:
  description: Simple API for USE sizer
  version: 1.2.0
  title: USE
  contact:
    email: DL-VTAS-VUPC-SERVICE@veritas.com
tags:
  - name: sizing
    description: Core Sizing
  - name: slps
    description: Storage Lifecycle Policies
  - name: workload-type
    description: Default Workload Attributes
  - name: skus
    description: Appliance SKUs
  - name: safety-margins
    description: Appliance Safety Margins
paths:
  /size:
    post:
      tags:
        - sizing
      summary: Run sizing
      description: ""
      requestBody:
        description: Sizing request
        required: true
        content:
          application/json:
            schema:
              $ref: "schemas/size_request.json"
      responses:
        "200":
          description: Successful sizing
          content:
            "application/json":
              schema:
                $ref: "schemas/size_response.json#/$defs/success"
        "405":
          description: Invalid input
          content:
            "application/json":
              schema:
                $ref: "schemas/util.json#/$defs/error"
  /async_size:
    post:
      tags:
        - sizing
      summary: Start asyncronous sizing request
      description: ""
      requestBody:
        description: Sizing request
        required: true
        content:
          application/json:
            schema:
              $ref: "schemas/size_request.json"
      responses:
        "200":
          description: Successful sizing
          content:
            "application/json":
              schema:
                $ref: "schemas/async_size_response.json"
        "405":
          description: Invalid input
          content:
            "application/json":
              schema:
                $ref: "schemas/util.json#/$defs/error"
  /slps:
    get:
      tags:
        - slps
      summary: List Storage Lifecycle Policies
      responses:
        "200":
          description: List of available SLPs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: schemas/slp.json
  "/slps/{name}":
    parameters:
      - in: path
        name: name
        required: true
        schema:
          type: string
    get:
      tags:
        - slps
      summary: Fetch a Storage Lifecycle Policy
      responses:
        "200":
          description: Definition of the named SLP
          content:
            application/json:
              schema:
                $ref: schemas/slp.json
        "404":
          description: Named policy does not exist
  /workload-type:
    get:
      tags:
        - workload-type
      summary: List Workload Types
      responses:
        "200":
          description: List of available Workload Types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: schemas/workload_type.json
  "/workload-type/{name}":
    parameters:
      - in: path
        name: name
        required: true
        schema:
          type: string
    get:
      tags:
        - workload-type
      summary: Fetch a Workload Type
      responses:
        "200":
          description: Definition of the named Workload Type
          content:
            application/json:
              schema:
                $ref: schemas/workload_type.json
        "404":
          description: Named workload type does not exist
  /skus:
    get:
      tags:
        - skus
      summary: List Appliance SKUs
      parameters:
        - in: query
          name: model
          required: false
          description: |
            Comma-separated list of models to include in result.
          schema:
            type: string
      responses:
        "200":
          description: List of available SKUs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplianceSku"
  /safety-margins:
    get:
      tags:
        - safety-margins
      summary: Fetch safety margins for all appliance models
      responses:
        "200":
          description: List of configured safety margins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplianceSafetyMargin"
  "/safety-margins/{model}":
    parameters:
      - in: path
        name: model
        required: true
        schema:
          type: string
    get:
      tags:
        - safety-margins
      summary: Fetch safety margins for specified appliance model
      responses:
        "200":
          description: |
            Configured safety margins for specified appliance model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplianceSafetyMargin"
        "404":
          description: Specified model has no configured safety margins
servers:
  - url: https://sizer.veritas.com/v1
components:
  schemas:
    ApplianceSafetyMargin:
      type: object
      properties:
        model:
          type: string
          description: |
            Appliance model to which these safety margins apply.
        capacity:
          type: number
          format: float
          description: |
            Maximum proportion of disk storage to use.  This must be a
            number between 0 and 1.
        cpu:
          type: number
          format: float
          description: |
            Maximum proportion of CPU time to use.  This must be a
            number between 0 and 1.
        memory:
          type: number
          format: float
          description: |
            Maximum proportion of appliance memory to use.  This must be
            a number between 0 and 1.
        network:
          type: number
          format: float
          description: |
            Maximum proportion of network bandwidth to use.  This must
            be a number between 0 and 1.
        io:
          type: number
          format: float
          description: |
            Maximum proportion of disk IO bandwidth to use.  This must
            be a number between 0 and 1.
        jobs_per_day:
          type: number
          description: |
            Maximum number of jobs that a single appliance should be
            allowed to do per day.
        databases:
          type: number
          description: |
            Maximum number of databases with 15-minute log backups that
            a single appliance should be allowed to backup.
        vms:
          type: number
          description: |
            Maximum number of virtual machines that a single appliance
            should be allowed to backup.
        max_streams:
          type: number
          description: |
            Limit on the maximum number of concurrent streams that an
            appliance can backup.
        files:
          type: number
          description: |
            Limit on the maximum number of files that an appliance can
            backup.
        msdp_pool:
          allOf:
            - $ref: schemas/slp.json
          description: |
            Maximum amount of disk storage that can be used for backup
            images.
    ApplianceSku:
      type: object
      properties:
        name:
          type: string
          description: Unique name for the SKU
        model:
          type: string
          enum:
            - "5150"
            - "5240"
            - "5250"
            - "5340"
            - 5340-HA
        shelves:
          type: number
          description: Number of shelves in the appliance configuration
        calculated_capacity:
          $ref: schemas/util.json#/$defs/size
        capacity:
          allOf:
            - $ref: schemas/util.json#/$defs/size
          description: Displayed storage capacity
        number_of_appliance_drives:
          type: number
          description: Number of appliance drives
        drives_per_shelf:
          type: number
          description: Number of drives per shelf
        number_of_shelf_drives:
          type: number
          description: Total number of drives in shelves
        number_of_total_drives:
          type: number
          description: Total number of drives
        number_of_calculated_drives:
          type: number
          description: Number of calculated drives
        drive_size:
          allOf:
            - $ref: schemas/util.json#/$defs/size
          description: Size of drives
        memory:
          allOf:
            - $ref: schemas/util.json#/$defs/size
          description: Available appliance memory
        io_config:
          type: string
          enum:
            - A
            - B
            - C
            - D
            - E
            - F
            - G
            - H
            - I
            - J
            - K
            - L
          description: IO configuration
        one_gbe:
          allOf:
            - $ref: "#/components/schemas/NetworkInterface"
          description: 1GbE network interfaces
        ten_gbe_copper:
          allOf:
            - $ref: "#/components/schemas/NetworkInterface"
          description: 10GbE network interfaces
        ten_gbe_sfp:
          allOf:
            - $ref: "#/components/schemas/NetworkInterface"
          description: 10GbE SFP interfaces
        twentyfive_gbe_sfp:
          allOf:
            - $ref: "#/components/schemas/NetworkInterface"
          description: 25GbE SFP interfaces
        eight_gbfc:
          allOf:
            - $ref: "#/components/schemas/NetworkInterface"
          description: 8Gbps FC interfaces
        sixteen_gbfc:
          allOf:
            - $ref: "#/components/schemas/NetworkInterface"
          description: 16Gbps FC interfaces
    NetworkInterface:
      type: object
      properties:
        count:
          type: number
          description: Number of interfaces
        io:
          type: number
          description: Total available bandwidth in MB/s
    SkuChange:
      type: object
      properties:
        add:
          type: array
          items:
            $ref: "#/components/schemas/ApplianceSku"
          description: List of new SKUs to add
        remove:
          type: array
          items:
            type: string
          description: List of SKU names to remove
    SizingResponse:
      type: array
      items:
        $ref: "#/components/schemas/SiteSizingResponse"
      description: |
        List of sizing results for per site.
    SiteSizingResponse:
      type: object
      properties:
        site_name:
          type: string
          description: |
            Name of the site which this sizing response is associated
            with.
        appliance_config:
          type: string
          description: |
            Appliance SKU selected for the site.
        appliances:
          type: array
          items:
            $ref: "#/components/schemas/SiteAppliances"
          description: |
            List of appliances assigned to the site.
    SiteAppliances:
      type: object
      properties:
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/ApplianceAssignments"
          description: |
            List of workloads assigned to this appliance, along with the
            number of clients of that workload.
        utilization:
          $ref: "#/components/schemas/ResourceUtilization"
    ApplianceAssignments:
      type: object
      properties:
        workload:
          type: string
          description: |
            Name of workload assigned.
        mode:
          type: string
          enum:
            - primary
            - dr
          description: |
            Whether this assignment is for the primary site of the
            workload, or a DR site for the workload.
        number_of_clients:
          type: number
          description: |
            Number of clients out of the total in the workload.
    ResourceUtilization:
      type: object
      description: |
        Year-over-year resource utilization for this appliance.  Each
        key refers to a list containing year-over-year utilization for
        that resource.  First element in the list is utilization for
        year 1, and so on.  There will be as many elements in the list
        as there are years being sized for.
      properties:
        capacity:
          type: array
          items:
            type: number
            format: double
          example:
            - 0.4
            - 0.5
            - 0.6
            - 0.7
            - 0.8
        cpu:
          type: array
          items:
            type: number
            format: double
          example:
            - 0.6
            - 0.6
            - 0.7
            - 0.7
            - 0.8
        memory:
          type: array
          items:
            type: number
            format: double
          example:
            - 0.5
            - 0.5
            - 0.5
            - 0.5
            - 0.5
        io:
          type: array
          items:
            type: number
            format: double
          example:
            - 0.2
            - 0.2
            - 0.2
            - 0.3
            - 0.3
        network:
          type: array
          items:
            type: number
            format: double
          example:
            - 0.1
            - 0.1
            - 0.1
            - 0.1
            - 0.1
    SitePref:
      type: object
      properties:
        site_name:
          type: string
          description: |
            Name of site for which these preferences are specified.
        appliance_model:
          type: string
          enum:
            - "5150"
            - "5240"
            - "5250"
            - "5340"
          description: |
            Choice of appliance model.
        appliance_config:
          type: string
          description: |
            Choice of specific appliance configuration.  If this is
            specified, automatic appliance selection is disabled.